requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pandas==2.1.4
numpy==1.26.3
sentence-transformers==2.3.1
scikit-learn==1.4.0
psycopg2-binary==2.9.9
nltk==3.8.1
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
transformers==4.36.2
torch==2.1.2
